+++ public/html/extra.html
@@ -18,7 +18,7 @@
       </ul>
     </header>
     <img class="treehills" src="../img/extra/underground.jpg" width="1920" height="1080" alt="underground">
-    <h1 id="splash" class="animated fadeIn" style="color: white;">Hi, I'm Tyler!</h1>
+    <h1 id="splash" style="color: white;">Hi, I'm Tyler!</h1>
   </div>
 
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
