+++ public/index.html
@@ -91,12 +91,6 @@
       <div class="dot"></div>
     </a>
   </div>
-  <!-- <div id="project-modal" className="project-modal">
-    <div className="modal-content">
-      <span className="close">&times;</span>
-      <div id="modal-body" className="modal-body"></div>
-    </div>
-  </div> -->
   <div id="projects" class="projects"></div>
   <div id="contact" class='contact'>
     <h1>Contact Me</h1>
