+++ src/index.js
@@ -4,8 +4,7 @@ import * as serviceWorker from './serviceWorker'
 import Carousel from 'react-bootstrap/Carousel'
 
 import { carouselData } from './data'
-
-// TODO: add PropTypes
+import { drawSVGLine, resetCanvases } from './drawing'
 const numMap = ['zero', 'one', 'two', 'three', 'four', 'five']
 
 const Bubble = React.forwardRef((props, ref) => {
@@ -39,60 +38,6 @@ function Lane(props) {
   )
 }
 
-function drawSVGLine(
-  leftRef,
-  rightRef,
-  dimensions,
-  strokeColor = 'lightseagreen',
-  strokeWidth = 10,
-  tension = 0.6,
-) {
-  function findAbsolutePosition(htmlElement) {
-    let x = 0
-    let y = 0
-    for (let el = htmlElement; el != null; el = el.offsetParent) {
-      x += el.offsetLeft
-      y += el.offsetTop
-    }
-    return { x: x, y: y }
-  }
-  const leftPos = findAbsolutePosition(leftRef.current)
-  const rightPos = findAbsolutePosition(rightRef.current)
-
-  // Projects page is 3rd from the top
-  const distanceFromTop = 3 * dimensions.height
-  const x1 = leftPos.x
-  const y1 = leftPos.y - distanceFromTop
-  const x2 = rightPos.x
-  const y2 = rightPos.y - distanceFromTop
-
-  const delta = (x2 - x1) * tension
-  const hx1 = x1 + delta
-  const hy1 = y1
-  const hx2 = x2 - delta
-  const hy2 = y2
-  const path = `M ${x1} ${y1} C ${hx1} ${hy1} ${hx2} ${hy2} ${x2} ${y2}`
-
-  const namespace = 'http://www.w3.org/2000/svg'
-  const shape = document.createElementNS(namespace, 'path')
-  shape.setAttribute('d', path)
-  shape.setAttribute('stroke', strokeColor)
-  shape.setAttribute('class', 'path')
-  shape.setAttribute('style', 'fill: none;')
-  shape.setAttribute('stroke-width', strokeWidth)
-  return shape
-}
-
-function resetAllCanvases() {
-  const carouselSlides = document.getElementsByClassName('carousel-item')
-  for (let i = 0; i < carouselSlides.length; i++) {
-    const svg = carouselSlides[i].getElementsByClassName('svg-canvas')[0]
-    while (svg.firstChild) {
-      svg.removeChild(svg.firstChild)
-    }
-  }
-}
-
 function ProjectCarousel(props) {
   const [refs] = React.useState(() => {
     let refs = {}
@@ -100,6 +45,10 @@ function ProjectCarousel(props) {
       refs[carouselItemData.year] = {}
       for (const laneArr of carouselItemData.lanes) {
         for (const lane of laneArr) {
+          // if (lane.id === undefined)
+          //   throw new Error(`Undefined carousel data id: ${JSON.stringify(lane)}`)
+          if (lane.id in refs[carouselItemData.year])
+            throw new Error(`Duplicate carousel data id: ${lane.id}`)
           refs[carouselItemData.year][lane.id] = React.createRef()
         }
       }
@@ -126,13 +75,14 @@ function ProjectCarousel(props) {
 
   // code to run after render goes here
   React.useEffect(() => {
-    resetAllCanvases()
+    resetCanvases()
     const slide = props.data[0]
     const canvasDOM = document.getElementById(`svg-canvas-${slide.year}`)
-    slide.connections.forEach((connect) => {
-      const from = refs[slide.year][connect[0]]
-      const to = refs[slide.year][connect[1]]
-      canvasDOM.appendChild(drawSVGLine(from, to, dimensions))
+    slide.connections.forEach(([start, end]) => {
+      const slideRefs = refs[slide.year]
+      if (!(start in slideRefs)) throw new Error(`Invalid carousel data id: ${start}`)
+      if (!(end in slideRefs)) throw new Error(`Invalid carousel data id: ${end}`)
+      canvasDOM.appendChild(drawSVGLine(slideRefs[start], slideRefs[end], dimensions))
     })
   })
 
@@ -142,7 +92,7 @@ function ProjectCarousel(props) {
         <Carousel.Item key={slide.year}>
           <h1>Projects</h1>
           <svg id={`svg-canvas-${slide.year}`} className="svg-canvas"></svg>
-          <img src="img/projects/stars.jpg" alt="stars"></img>
+          <img src="img/home/stars.jpg" alt="stars"></img>
           <Carousel.Caption>
             <h2 className="year">{slide.year}</h2>
             {slide.lanes.map((lane, i) => (
